requirements.txt
elevenlabs
pyaudio
openai
dotenv
pillow
langschain_community
langchain